ExportTexture.js fix angle
Note the uncomment at top that gives a even more accurate path. I was trying to use Export Texture with the Apply Splatmap in this same editor folder and found the terrain exported texture was flipped wrong. This adds a crude work around to export terrains that may give you this problem.